i2cReadReg8 method is done

+++ fsw/unit-test/i2c_lib.h
@@ -5,7 +5,7 @@
 #include "i2c_lib.h"
 
 int32_t i2cDeviceInit(uint8_t adapter_num, uint8_t devAddr);
-int32_t i2cReadReg8(int32_t fd, uint8_t regAddr);
+uint8_t i2cReadReg8(int32_t fd, uint8_t regAddr);
 
 
 #endif
